Task 6

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,12 @@ int main(void)
 	long int d;
 	long long int e;
 	float f;
-	printf("Size of a char: %lu byts(s)\n", (unsigned int)sizeof(a));
-	printf("Size of an int: %lu byts(s)\n", (unsigned int)sizeof(b));
-	printf("Size of a double: %lu byts(s)\n", (unsigned int)sizeof(c));
-	printf("Size of a long int: %lu byts(s)\n", (unsigned int)sizeof(d));
-	printf("Size of a long long int: %lu byts(s)\n", (unsigned int)sizeof(e));
-	printf("Size of a float: %lu byts(s)\n", (unsigned int)sizeof(f));
+
+	printf("Size of a char: %lu byts(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu byts(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a double: %lu byts(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long int: %lu byts(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a long long int: %lu byts(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a float: %lu byts(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
